requirements.txt
Flask
psycopg[binary]
python-dotenv==0.21.0
Flask-Cors==3.0.10
gunicorn==20.1.0
flasgger==0.9.7.1
PyYAML>=3.0
jsonschema>=3.0.1